ResultModal: Extract out `useMatcher` hook
